refactor: decompose CommandLineValidator into Parser, Validator, Builder (fixes #178)
Split the 1332-line CommandLineValidator god class into 5 separate modules:
- CliArgument: metadata-driven argument definitions
- CliParser: parses CLI args into ParsedArguments (extracted from ParseArguments)
- CliValidator: cross-argument validation rules (extracted from ValidateRequired/ValidateOptional)
- RequestBuilder: maps parsed+validated args to FileGenerationRequest (extracted from CreateFileGenerationRequest)
- HelpTextGenerator: auto-generated help from argument metadata (extracted from ShowUsage)

CommandLineValidator is now a thin ~35-line facade that delegates to the new Cli modules.

Benefits:
- Each module has a small interface with high leverage behind it
- Parse, validate, and build can be tested independently
- Adding a flag now touches 1-2 modules instead of 6
- Help text is self-documenting via CliArgument metadata
